+++ CHANGELOG.md
@@ -52,6 +52,8 @@ The Koto project adheres to
     exception being thrown.
 - Objects can be compared with `null` on the LHS without having to implement 
   `KotoObject::equal` and/or `not_equal`.
+- `map.with_meta_map` has been replaced with `set_meta`, and `get_meta_map` has
+  been renamed to `get_meta`.
 
 #### API